Centralize gateway policy and type storage queries

+++ crates/storage/src/request_metadata.rs
@@ -20,8 +20,8 @@ pub use reconciliation::{
     RequestMetadataGatewayEpochState, RequestMetadataLossReport,
 };
 
-pub(crate) const REQUEST_METADATA_EVENT_REPLAY_HORIZON_DAYS: i64 = 7;
-pub(crate) const REQUEST_METADATA_EVENT_FUTURE_SKEW_MINUTES: i64 = 5;
+pub(crate) const REQUEST_METADATA_EVENT_REPLAY_HORIZON_DAYS: i32 = 7;
+pub(crate) const REQUEST_METADATA_EVENT_FUTURE_SKEW_MINUTES: i32 = 5;
 
 #[cfg(test)]
 use chrono::Utc;
